requirements.txt
openai
python-dotenv
pydub
yt-dlp